moving topvisitlogs fn to visitlogcarservicefile
