+++ src/component/EventCardService.js
@@ -185,7 +185,7 @@ export const fetchOfficialEvents = async () => {
   }
 };
 
-const fetchUserName = async (uid) => {
+export const fetchUserName = async (uid) => {
   // Reference to the uid instead of the docid of the user.
   const userQuery = query(
     collection(db, USERS_COLLECTION),
@@ -649,37 +649,6 @@ export const fetchUserOutreaches = async () => {
     logEvent(
       "STREET_CARE_ERROR",
       `error on fetchUserOutreaches in EventCardService.js- ${error.message}`
-    );
-    throw error;
-  }
-};
-
-export const fetchTopVisitLogs= async () => {
-  try {
-    const visitlogs = collection(db, PERSONAL_VISIT_LOG);
-    const visitlogsQuery = query(
-      visitlogs, 
-      orderBy('dateTime', 'desc'), // Order visit logs by the 'dateTime' field in descending order to get the newest entries first
-      limit(6) // Limit to top 6 records
-    );
-    const visitLogDocRef = await getDocs(visitlogsQuery);
-    let visitLogs = [];
-    for (const doc of visitLogDocRef.docs) {
-      const visitLogData = doc.data(); 
-      const id = doc.id;
-      const userName = await fetchUserName(visitLogData.uid);
-      visitLogs.push({
-        ...visitLogData,
-        userName: userName,
-        id: id,
-      });
-    }
-    console.log(visitLogs)
-    return visitLogs;
-  } catch (error) {
-    logEvent(
-      "STREET_CARE_ERROR",
-      `error on fetchTopVisitLogs EventCardService.js- ${error.message}`
     );
     throw error;
   }
